Refactor Upload class
